feat: validate bookmark files
The JSON structure supported by this extension was never explictly documented
and only explained by examples. Due to upcoming features and knowing how manipulating
(deeply) nested JSON by hand is error-prone it makes sense to perform validation
before attempting to import the bookmark files.

I defined a JSON schema and additionally published it to a public location.
JSON Schema for Bookmarks:
https://frederikb.github.io/bookmarksync/schemas/bookmarks.1-0-0.schema.json
In the (rare) case that the schema is updated we need to ensure that the new version is published
as well.
The schema is versioned according to SchemaVer.

As a side benefit, users may also reference the schema in their bookmark JSON files,
via a $schema entry on the root and gain immediate validation in their editor (if supported).

Fixes #29