feat: check configured bookmark connection
Give the user a button in the options page to check the configured
connection. This enables him to check the connection and receive
detailed error messages without having to force a sync.

Additionally improved the level of detail of error messages to better
support the user with various misconfiguration scenarios.

Closes #52