fix: inconsistent order of bookmarks (#72)
Bookmarks were added in a non-deterministic order which could change
each time. This was due to the creation of all bookmarks and folders
happening in parallel.

The new approach forces creation to happen in order.

This will likely impact the order of the bookmarks for most use cases,
but then should be stable and follow the order specified in the input
JSON files.